Half-way done, no testing yet

+++ Server/Tasks/Domain/DbEntities/Project.cs
@@ -10,14 +10,14 @@ namespace Core.Domain.DbEntities
             Tasks = new HashSet<Tasks>();
         }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
         public DateTime? CreatedDate { get; set; }
-        public int? CreatedBy { get; set; }
+        public long? CreatedBy { get; set; }
         public DateTime? UpdatedDate { get; set; }
-        public int? UpdatedBy { get; set; }
+        public long? UpdatedBy { get; set; }
         public bool Deleted { get; set; }
 
         public virtual Project Parent { get; set; }
